tps/requirements.txt
celery==4.0.2
Django==1.9.7
django-allauth==0.25.2
django-bootstrap3==7.1.0
git+git://github.com/mohammadroghani/django-clone.git
django-multiselectfield==0.1.4
mock==2.0.0
model-mommy==1.3.0
django-import-export
gnupg
django-debug-toolbar
redis
python-magic
pygit2
django-extensions
ansi2html==1.2.0
django-redis